+++ schema/timestamptz.go
@@ -22,7 +22,7 @@ const (
 )
 
 type TimestamptzTransformer interface {
-	TransformTimestamptz(*Timestamptz) interface{}
+	TransformTimestamptz(*Timestamptz) any
 }
 
 type Timestamptz struct {
@@ -41,7 +41,10 @@ func (dst *Timestamptz) Equal(src CQType) bool {
 	}
 
 	if value, ok := src.(*Timestamptz); ok {
-		return dst.Status == value.Status && dst.Time.Equal(value.Time) && dst.InfinityModifier == value.InfinityModifier
+		if dst.Status != value.Status || dst.InfinityModifier != value.InfinityModifier {
+			return false
+		}
+		return dst.Time.Equal(value.Time)
 	}
 
 	return false
@@ -55,13 +58,13 @@ func (dst *Timestamptz) String() string {
 	}
 }
 
-func (dst *Timestamptz) Set(src interface{}) error {
+func (dst *Timestamptz) Set(src any) error {
 	if src == nil {
 		*dst = Timestamptz{Status: Null}
 		return nil
 	}
 
-	if value, ok := src.(interface{ Get() interface{} }); ok {
+	if value, ok := src.(interface{ Get() any }); ok {
 		value2 := value.Get()
 		if value2 != value {
 			return dst.Set(value2)
@@ -108,7 +111,7 @@ func (dst *Timestamptz) Set(src interface{}) error {
 	return nil
 }
 
-func (dst Timestamptz) Get() interface{} {
+func (dst Timestamptz) Get() any {
 	switch dst.Status {
 	case Present:
 		if dst.InfinityModifier != None {
@@ -123,7 +126,7 @@ func (dst Timestamptz) Get() interface{} {
 }
 
 func (dst *Timestamptz) DecodeText(src []byte) error {
-	if src == nil {
+	if len(src) == 0 {
 		*dst = Timestamptz{Status: Null}
 		return nil
 	}